fix(build): fixes 21 assorted issues
PR #5199

+++ code/game/objects/structures/stool_bed_chair_nest/bed.dm
@@ -232,15 +232,16 @@
 	return // Doesn't care about material or anything else.
 
 /obj/structure/bed/roller/attackby(obj/item/weapon/W as obj, mob/user as mob)
-	if(isWrench(W) || istype(W,/obj/item/stack) || isWirecutter(W))
+	if(isWrench(W) || istype(W, /obj/item/stack) || isWirecutter(W))
 		return
-	else if(istype(W,/obj/item/roller_holder))
+	else if(istype(W, /obj/item/roller_holder))
 		if(buckled_mob)
 			user_unbuckle_mob(user)
-		else
+		else if(rollertype)
 			visible_message("[user] collapses \the [src.name].")
 			new rollertype(get_turf(src))
-			QDEL_IN(src, 0)
+			rollertype = null
+			qdel(src)
 		return
 	..()
 
@@ -264,17 +265,19 @@
 	bedtype = /obj/structure/bed/roller/adv
 
 /obj/item/roller/attack_self(mob/user)
+	if(!bedtype)
+		return
 	var/obj/structure/bed/roller/R = new bedtype(user.loc)
 	R.add_fingerprint(user)
+	bedtype = null
 	qdel(src)
 
-/obj/item/roller/attackby(obj/item/weapon/W as obj, mob/user as mob)
-
-	if(istype(W,/obj/item/roller_holder))
+/obj/item/roller/attackby(obj/item/weapon/W, mob/user)
+	if(istype(W, /obj/item/roller_holder))
 		var/obj/item/roller_holder/RH = W
 		if(!RH.held)
 			to_chat(user, "<span class='notice'>You collect the roller bed.</span>")
-			src.forceMove(RH)
+			forceMove(RH)
 			RH.held = src
 			return
 	..()
@@ -290,8 +293,7 @@
 	..()
 	held = new /obj/item/roller(src)
 
-/obj/item/roller_holder/attack_self(mob/user as mob)
-
+/obj/item/roller_holder/attack_self(mob/user)
 	if(!held)
 		to_chat(user, "<span class='notice'>The rack is empty.</span>")
 		return
@@ -299,10 +301,9 @@
 	to_chat(user, "<span class='notice'>You deploy the roller bed.</span>")
 	var/obj/structure/bed/roller/R = new held.bedtype(user.loc)
 	R.add_fingerprint(user)
-	qdel(held)
-	held = null
+	QDEL_NULL(held)
 
-/obj/structure/bed/roller/post_buckle_mob(mob/living/M as mob)
+/obj/structure/bed/roller/post_buckle_mob(mob/living/M)
 	if(M == buckled_mob)
 		set_density(1)
 		icon_state = "[initial(icon_state)]_up"
@@ -315,11 +316,16 @@
 /obj/structure/bed/roller/MouseDrop(over_object, src_location, over_location)
 	..()
 	if((over_object == usr && (in_range(src, usr) || usr.contents.Find(src))))
-		if(!ishuman(usr))	return
-		if(buckled_mob)	return 0
+		if(!rollertype)
+			return
+		if(!ishuman(usr))
+			return
+		if(buckled_mob)
+			return 0
 		visible_message("[usr] collapses \the [src.name].")
 		new rollertype(get_turf(src))
-		QDEL_IN(src, 0)
+		rollertype = null
+		qdel(src)
 		return
 
 ///
